==> 0_news_crawling.ipynb <==
import pandas as pd
import numpy as np

from bs4 import BeautifulSoup
import requests

import time
import re
import selenium
from selenium import webdriver as wd
from selenium.webdriver import Chrome
from selenium.webdriver.common.keys import Keys

from tqdm import tqdm_notebook as tqdm

# 크롤링해야 할 정보

- 기사 헤드라인
- 기사 주소 url
- 업로드 날짜, 시간
- 신문사명
- 본문 내용
- 최대 4페이지 까지 (try구문 사용!)

## 2021 날짜 정보
- 31: 1,3, 5, 7, 8, 10, 12
- 30: 4, 6, 9, 11
- 28: 2

day=[str(i).zfill(2) for i in range(1,32)] #### User-Change ####

day

#### User-Change ####
url_info="https://news.naver.com/main/list.naver?mode=LS2D&mid=shm&sid2=310&sid1=101&date=202112" #### User-Change ####
url_lst=[]
for d in day:
    url=url_info+str(d)
    for num in range(1,5):
        url_lst.append(url+'&page='+str(num))

url_lst
# https://news.naver.com/main/list.naver?mode=LS2D&mid=shm&sid2=310&sid1=101&date=20210101&page=2

len(url_lst)

### soup으로 다음과 같은 정보부터 가져오기
# 기사 헤드라인
# 기사 주소 url
# 업로드 날짜, 시간
# 신문사명



title=[]
ar_url=[]
day=[]
time_=[]
com=[]


for u in url_lst:
    print(u)
    page=requests.get(u,headers={'User-Agent':'Mozilla/5.0'})
    soup = BeautifulSoup(page.content, "html.parser")

    #### 상위 10개 기사 ####
    top10=soup.select('#main_content > div.list_body.newsflash_body > ul.type06_headline > li')
    for i in range(0,len(top10)):
        try:
            # title
            title.append(top10[i].find_all('a')[1].text.replace('\n','').replace('\t',''))
            # ar_url
            ar_url.append(top10[i].find_all('a')[1]['href'])
            # day
            day.append(top10[i].find_all('span','date')[0].text.split('. ')[0])
            #time
            time_.append(top10[i].find_all('span','date')[0].text.split('. ')[1])
            #com
            com.append(top10[i].find_all('span','writing')[0].text)
        except:
            print('Out of Range: ',i)
            # title
            title.append(top10[i].find_all('a')[0].text.replace('\n','').replace('\t',''))
            # ar_url
            ar_url.append(top10[i].find_all('a')[0]['href'])
            # day
            day.append(top10[i].find_all('span','date')[0].text.split('. ')[0])
            # time
            time_.append(top10[i].find_all('span','date')[0].text.split('. ')[1])
            # com
            com.append(top10[i].find_all('span','writing')[0].text)  

    #### 하위 10개 기사 ####
    top20=soup.select('#main_content > div.list_body.newsflash_body > ul.type06 > li')
    for i in range(0,len(top20)):
        try:
            # title
            title.append(top20[i].find_all('a')[1].text.replace('\n','').replace('\t',''))
            # ar_url
            ar_url.append(top20[i].find_all('a')[1]['href'])
            # day
            day.append(top20[i].find_all('span','date')[0].text.split('. ')[0])
            #time
            time_.append(top20[i].find_all('span','date')[0].text.split('. ')[1])
            #com
            com.append(top20[i].find_all('span','writing')[0].text)

        except:
            print('Out of Range: ',i)
            # title
            title.append(top20[i].find_all('a')[0].text.replace('\n','').replace('\t',''))
            # ar_url
            ar_url.append(top20[i].find_all('a')[0]['href'])
            # day
            day.append(top20[i].find_all('span','date')[0].text.split('. ')[0])
            # time
            time_.append(top20[i].find_all('span','date')[0].text.split('. ')[1])
            #com
            com.append(top20[i].find_all('span','writing')[0].text)
            

            

print(len(title))
print(len(ar_url))
print(len(day))
print(len(time_))
print(len(com))

month1=pd.DataFrame({'title':title,'url':ar_url,'day':day,'time':time_,'company':com})
month1

month1.isna().sum()

# contents

contents=[]

for u in tqdm(list(month1['url']),desc='tqdm'):
    #print(u)
    page=requests.get(u,headers={'User-Agent':'Mozilla/5.0'})
    soup = BeautifulSoup(page.content, "html.parser")
    try:
        contents.append(soup.find('div',{'class':'newsct_article _article_body'}).text)
    except:
        print(u)
        contents.append(np.nan)

print(len(title))
print(len(ar_url))
print(len(day))
print(len(time_))
print(len(com))
print(len(contents))

month1['contents']=contents

month1

month1.isna().sum()

month1.to_csv('month12.csv',encoding='utf-8',index=False)

month11=pd.read_csv('month11.csv')
month11.head(10)

month11.tail(10)

# merged

month=[i for i in range(1,13)]
month

mon1=pd.read_csv('month1.csv')
mon1.isna().sum()

df=pd.read_csv('month1.csv')

for i in month[1:]:
    m=pd.read_csv('month'+str(i)+'.csv')
    print(i)
    print(m.isna().sum())
    print('------------')
    df=pd.concat([df,m])

df

df.isna().sum()

mon1=pd.read_csv('month1.csv')
mon2=pd.read_csv('month2.csv')
mon3=pd.read_csv('month3.csv')
print(mon1.shape)
print(mon2.shape)
print(mon3.shape)
pd.concat([mon1,mon2])

2348+2159
