==> 4_XGBoost_upload.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

ls

import numpy as np
import matplotlib.pyplot as plt
import pandas as pd

import pickle
import xgboost
from sklearn.metrics import mean_squared_error

from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV

tr=pd.read_csv('data/train_fin_.csv')
train=tr

ts=pd.read_csv('data/test_fin_.csv')
test=ts

y_train=train['cltv']
x_train=train.drop('cltv',axis=1)

y_test=test['cltv']
x_test=test.drop('cltv',axis=1)

xgb_model=xgboost.XGBRegressor(n_estimators=100, learning_rate=0.08, gamma=0, subsample=0.75,
                           colsample_bytree=1, max_depth=7)
xgb_model.fit(x_train,y_train)

file_name='xgb_reg.pkl'
pickle.dump(xgb_model,open(file_name,'wb'))
#xgb_model.save_model('/model/XGBoost.model')

xgboost.plot_importance(xgb_model)

predictions = xgb_model.predict(x_test)
print(mean_squared_error(y_test, predictions)**0.5)
