==> 5_cluster_upload.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

ls

import pandas as pd
import numpy as np

# train=pd.read_csv('data/train_fin_.csv')
# test=pd.read_csv('data/test_fin_.csv')
# tot=pd.concat([train,test])
# tot.to_csv('tot_fin_.csv')

#!pip install tslearn

from sklearn.cluster import KMeans
from sklearn.decomposition import PCA

import matplotlib.pyplot as plt
import seaborn as sns
from sklearn import preprocessing
import pickle

df=pd.read_csv('data/tot_fin_.csv')
df.columns

df1=df.groupby(['cust']).agg({'hour':'mean','buy_am':'sum','buy_ct':'sum','time_diff':'sum',
                          'month':'mean', 'ages':'mean','month_diff':'mean',
                          'cop_diff':'mean','cm_diff':'mean','loc_diff':'mean','weekend':'sum',
                          'buy_level':'mean','lag_buy':'sum','lag_income':'sum','cltv':'sum', 'gender_female':'mean','gender_male':'mean',
                          'cop_A':'sum','cop_B':'sum','cop_C':'sum','cop_D':'sum','cop_E':'sum','channel_offline':'sum','channel_online':'sum',
                          'loc_Z05':'sum','loc_Z06':'sum', 'loc_Z07':'sum', 'loc_Z08':'sum', 'loc_Z09':'sum',
       'loc_Z10':'sum', 'loc_Z11':'sum', 'loc_Z12':'sum', 'loc_Z13':'sum',
       'loc_Z14':'sum', 'loc_Z15':'sum', 'loc_Z16':'sum', 'loc_Z17':'sum',
       'loc_resid_Z01':'sum', 'loc_resid_Z02':'sum', 'loc_resid_Z03':'sum',
       'loc_resid_Z04':'sum', 'loc_resid_Z05':'sum', 'loc_resid_Z06':'sum',
       'loc_resid_Z07':'sum', 'loc_resid_Z08':'sum', 'loc_resid_Z09':'sum',
       'loc_resid_Z10':'sum', 'loc_resid_Z11':'sum', 'loc_resid_Z12':'sum',
       'loc_resid_Z13':'sum', 'loc_resid_Z14':'sum', 'loc_resid_Z15':'sum',
       'loc_resid_Z16':'sum', 'loc_resid_Z17':'sum'
                          })
df1

#df=pd.read_csv('data/tot_fin_.csv')
#df=df.set_index('cust').reset_index(drop=True)
scaler=preprocessing.StandardScaler()
df_sc=scaler.fit_transform(df1)
df_sc.shape

kmean= KMeans(3)
kmean.fit(df_sc)
labels=kmean.labels_

file_name='kmean.pkl'
pickle.dump(kmean,open(file_name,'wb'))

cluster=pd.concat([df1,pd.DataFrame({'cluster':labels})],axis=1)
cluster.head()

dummy_var1=['loc_Z05', 'loc_Z06','loc_Z07', 'loc_Z08', 'loc_Z09', 'loc_Z10',
            'loc_Z11', 'loc_Z12', 'loc_Z13', 'loc_Z14',
            'loc_Z15', 'loc_Z16', 'loc_Z17']
dummy_var2=['loc_resid_Z01','loc_resid_Z02', 'loc_resid_Z03', 'loc_resid_Z04',
       'loc_resid_Z05', 'loc_resid_Z06', 'loc_resid_Z07',
       'loc_resid_Z08', 'loc_resid_Z09', 'loc_resid_Z10',
       'loc_resid_Z11', 'loc_resid_Z12', 'loc_resid_Z13',
       'loc_resid_Z14', 'loc_resid_Z15', 'loc_resid_Z16',
       'loc_resid_Z17'] 
dummy_var3=['cop_A', 'cop_B', 'cop_C', 'cop_D', 'cop_E']
dummy_var4=['channel_offline', 'channel_online']

cluster['loc']=cluster[dummy_var1].idxmax(axis=1).apply(lambda x:x[-3:])
cluster=cluster.drop(dummy_var1,axis=1)
print(cluster.shape)

cluster.columns

cluster['loc_resid']=cluster[dummy_var2].idxmax(axis=1)[-3:]
cluster=cluster.drop(dummy_var2,axis=1)
print(cluster.shape)

cluster['cop']=cluster[dummy_var3].idxmax(axis=1).apply(lambda x:x[-1])
cluster=cluster.drop(dummy_var3,axis=1)
print(cluster.shape)

cluster['channel']=cluster[dummy_var4].idxmax(axis=1)
cluster=cluster.drop(dummy_var4,axis=1)
print(cluster.shape)

cluster.columns

cluster['cluster'].value_counts()

c0=cluster[cluster['cluster']==0]
c0

sns.set(style='ticks',color_codes=True)
g=sns.pairplot(c0)
plt.show()

c1=cluster[cluster['cluster']==1]
sns.set(style='ticks',color_codes=True)
g=sns.pairplot(c1)
plt.show()

c2=cluster[cluster['cluster']==2]
sns.set(style='ticks',color_codes=True)
g=sns.pairplot(c2)
plt.show()

plt.hist(np.log(c0['cltv']))

c0['cop'].value_counts()

c1=cluster[cluster['cluster']==1]
plt.hist(np.log(c1['cltv']))

c2=cluster[cluster['cluster']==2]
plt.hist(np.log(c2['cltv']))

c2.columns

plt.hist(np.log(c0['lag_buy']))

plt.hist(np.log(c1['lag_buy']))

plt.hist(np.log(c2['lag_buy']))

plt.scatter(c0['hour'],c0['buy_am'])

plt.scatter(c1['hour'],c1['buy_am'])

plt.scatter(c2['hour'],c2['buy_am'])

plt.figure(0)
plt.hist(c0['buy_level'])
plt.figure(1)
plt.hist(c1['buy_level'])
plt.figure(2)
plt.hist(c2['buy_level'])
#print(c1['buy_level'].value_counts())
#print(c2['buy_level'].value_counts())

plt.figure(0)
plt.hist(np.log(c0['lag_income']))
plt.figure(1)
plt.hist(np.log(c1['lag_income']))
plt.figure(2)
plt.hist(np.log(c2['lag_income']))

plt.figure(0)
plt.hist(np.log(c0['lag_buy']))
plt.figure(1)
plt.hist(np.log(c1['lag_buy']))
plt.figure(2)
plt.hist(np.log(c2['lag_buy']))

plt.figure(0)
plt.hist(c0['month'])
plt.figure(1)
plt.hist(c1['month'])
plt.figure(2)
plt.hist(c2['month'])
# 클러스터에 관계 x

plt.figure(0)
plt.scatter(c0['month'],c0['buy_am'])
plt.figure(1)
plt.scatter(c1['month'],c1['buy_am'])
plt.figure(2)
plt.scatter(c2['month'],c2['buy_am'])
